Split up window and process
